Struct f3::i2c::I2c

pub struct I2c<S> { /* fields omitted */ }

I2C

Methods

impl I2c<Free>

fn new() -> Option<I2c<Free>>

Initializes the I2C peripheral

NOTE There can only be an instance of this peripheral. If you call this constructor a second time, it will return None.

fn read(self, address: Address) -> Read

Reads a single byte from a slave with a certain address

fn read_exact<B>(self, address: Address, buffer: B) -> ReadExact<B> where B: AsMut<[u8]>

Reads bytes from a slave with a certain address into a buffer

fn write(self, address: Address, byte: u8) -> Write

Sends a single byte to a slave with a certain address

fn write_all<B>(self, address: Address, bytes: B) -> WriteAll<B> where B: AsRef<[u8]>

Sends several bytes to a slave with a certain address

impl I2c<Busy>

fn flush(self) -> Flush

Waits until the current transmission ends

fn read(self, address: Option<Address>) -> Read

Reads a single byte from a slave into buffer

address = None, to continue communicating with the same slave

address = Some(_), to switch to another slave without releasing the bus

fn read_exact<B>(self, address: Option<Address>, buffer: B) -> ReadExact<B> where B: AsMut<[u8]>

Reads several bytes from a slave into buffer

address = None, to continue communicating with the same slave

address = Some(_), to switch to another slave without releasing the bus

fn write(self, address: Option<Address>, byte: u8) -> Write

Sends a single byte to a slave

address = None, to continue communicating with the same slave

address = Some(_), to switch to another slave without releasing the bus

fn write_all<B>(self, address: Option<Address>, bytes: B) -> WriteAll<B> where B: AsRef<[u8]>

Sends bytes to a slave

address = None, to continue communicating with the same slave

address = Some(_), to switch to another slave without releasing the bus

fn stop(self) -> Stop

Ends the I2C transmission and frees the bus