Struct f3::spi::Spi

pub struct Spi { /* fields omitted */ }

SPI

Methods

impl Spi

fn new() -> Option<Spi>

Initializes the SPI peripheral

NOTE There can only be an instance of this peripheral. If you call this constructor a second time, it will return None.

fn transfer(self, byte: u8) -> Transfer

Sends one byte to the slave and returns the byte that the slave returns

fn transfer_all<B>(self, bytes: B) -> TransferAll<B> where B: AsMut<[u8]> + AsRef<[u8]>

Sends bytes to the slave and returns the bytes that the slave returned