Struct f3::timer::Timer

pub struct Timer { /* fields omitted */ }

A timer

Methods

impl Timer

fn new() -> Option<Self>

Initializes the Timer peripheral

NOTE There can only be an instance of this peripheral. If you call this constructor a second time, it will return None.

fn oneshot(self, ms: u16) -> OneShot

Sets a one time "alarm"

fn periodic(&mut self, ms: u16) -> Periodic

Sets a periodic "alarm"